fix: fixed location of the code
