+++ app/src/main/java/com/github/cstrerath/uncover/QuestStepDao.kt
@@ -1,18 +1,17 @@
 package com.github.cstrerath.uncover
 
-import androidx.room.Entity
-import androidx.room.PrimaryKey
+import androidx.room.Dao
+import androidx.room.Insert
+import androidx.room.Query
 
-enum class StepType {
-    INITIAL, SOLUTION, COMPLETION
+@Dao
+interface QuestStepDao {
+    @Query("SELECT * FROM quest_steps WHERE questId = :questId")
+    fun getStepsForQuest(questId: Int): List<QuestStep>
+
+    @Query("SELECT * FROM quest_steps WHERE questId = :questId AND stepType = :stepType")
+    fun getStepForQuestAndType(questId: Int, stepType: StepType): QuestStep
+
+    @Insert
+    fun insertQuestStep(questStep: QuestStep)
 }
-
-@Entity(tableName = "quest_steps")
-data class QuestStep(
-    @PrimaryKey val stepId: Int,
-    val questId: Int,
-    val stepType: StepType,
-    val warriorVariantKey: String,
-    val thiefVariantKey: String,
-    val mageVariantKey: String
-)
